Large changes to AREPO frontend accomodate the new halo/subhalo/(in ray files) line of sight metadata, revamp the unit system to be more comprehensive, and to facilitate using Trident. Other changes include giving units to spectral properties determined by Trident and defining convenience functions (gas(), stars(), bh(), los()) that make accessing data more uniform. To do this we use convenience funcions like gas() and hs(), which read the gas particle data and halo/subhalo metadata.
gas() and stars() use the standard yt ds.r['PartTypeN', 'Property'] syntax at their core, but the gas() function is complicated because it has to be defined slightly differently for halo vs ray files; we want to be able to access spectral properties derived by Trident; and because gas properties are "layered" with 2 snapshots in the same HDF5 file, giving vector properties multiple orientations. These functions also handle any last-minute unit corrections.

hs() and los() are used to load metadata which is usually not accessible in yt. So we have to construct data objects to hold this new info. This is done using _get_hsvals() and _get_rayvals() methods, which give properties correct units and vector properties the correct unit system.